Pull out reusable image input functions
